Move responders to query package.
