+++ src/main/java/com/lockarhythm/query/QueryInterpreter.java
@@ -1,7 +1,5 @@
 package com.lockarhythm.query;
 
-import com.lockarhythm.responders.Response;
-
 public interface QueryInterpreter {
   public Response interpret(String query); // TODO: rename to QueryResult
 }
