+++ src/main/java/com/lockarhythm/query/QueryRespondable.java
@@ -1,4 +1,4 @@
-package com.lockarhythm.responders;
+package com.lockarhythm.query;
 
 public interface QueryRespondable {
   Response respondTo(String query);
