+++ src/main/java/com/lockarhythm/query/echo/EchoResponder.java
@@ -1,7 +1,7 @@
-package com.lockarhythm.responders.echo;
+package com.lockarhythm.query.echo;
 
-import com.lockarhythm.responders.QueryRespondable;
-import com.lockarhythm.responders.Response;
+import com.lockarhythm.query.QueryRespondable;
+import com.lockarhythm.query.Response;
 
 public class EchoResponder implements QueryRespondable {
   public Response respondTo(String query) {
